Update to 1.3.0.
1.3.0:
This release includes a good deal of refactoring and cleanup, along with
bug fixes and build configuration improvements.   Thanks to the wide range
of contributors represented in the changelog below.

1.2.0:
This minor feature release adds the -query option to print the current
RLMVO layout in use by the X server.