Update to 0.11.0, requested by Patrick Welche. Bump dependency in buildlink3.mk since its only user is dconf-editor. Set LICENSE. Add comments to patches.
Changes in dconf 0.10.0
=======================

The version number was increased and a new entry was added to the NEWS.

Changes in dconf 0.9.1
======================

  - give a g_warning() on failure to communicate with service

  - remove unworking 'set lock' call from dconf API and commandline tool

  - add code to exit gracefully on receipt of SIGINT, SIGHUP, SIGTERM

  - remove "service function" logic; always use the XDG runtime directory

Changes in dconf 0.9.0
======================

There has been an extremely minor incompatible change in the D-Bus API
of dconf this release.  From a practical standpoint, this change will
have no effect.  However, it serves as a reminder that the dconf D-Bus
API is private and can change from version to version (and will likely
change in the future).

As such, it is appropriate for those packaging dconf to kill all running
instances of dconf ('killall dconf-service') as part of their postinst
for the package.  It will be dbus-activated again on the next use.

  - support loading/storing of maybe types in dconf

  - remove NFS detection hackery and rely on XDG runtime dir

  - add proper support for change notification to DConfClient

  - commandline tool improvements

    - reset: reset keys or entire subpaths

    - dump: dump entire subpaths to keyfile format

    - load: load them back again (maybe at a different path)

    - watch: actually works now

  - editor improvements

    - keys now change in editor when changed from outside

    - support for flags

    - show dconf-editor in applications list

  - work around incompatible Vala bindings changes with an #if

  - don't install the bash completion script as executable

  - fix a warning caused by reusing a GError variable

  - other small fixes


Changes in dconf 0.7.5
======================

This release corrects a serious flaw in the previous release: crashing
if the database did not already exist.

It also contains many fixes and improvements to the dconf-editor,
including use of GSettings to store the window geometry.

This is the final release before 0.8.0 which will become the first
release in a new stable series.  Feature development will continue on
'master' toward 0.9 past that point.

Changes in dconf 0.7.4
======================

Changes in this version:

  - #648949: multithreading issue fixed (which actually affects all
    GSettings-using programs since dconf is used from a helper thread in
    that case)

  - dconf commandline tool is vastly more friendly now

    - no more aborting on unrecognised arguments

    - proper help

    - bash completion support

  - support for sysadmin lockdown

  - the editor now properly reads installed enum xml files

Changes in dconf 0.7.3
======================

This release consists almost entirely of fixes made by Robert to
dconf-editor.  A few other trivial build fixes are included as well
(bumping library version dependencies to match reality, etc).

Changes in dconf 0.7.2
======================

This is entirely a cleanup/fixes release.  Some fixes here to make the
increasingly-strict toolchain happy, and also some fixes for some
crashers in the GSettings backend and service.

 - remove some unused variables (new GCC gives a warning: #640566, another)
 - add a mutex to fix multi-threading issue (#640611)
 - don't crash if we have no D-Bus
 - clean up symbol exports
 - fix a crash in the service when using 'reset'
 - drop old linker options that were for libtool

Changes in dconf 0.7.1
======================

The last release contained a few problems that caused build failures on
some strict linkers.  Those should be fixed now.

Changes in dconf 0.7
====================

 - new library to use dconf with libdbus-1
 - quite a lot of improvements and bug-fixes in dconf-editor, thanks to
   Robert Ancell
 - some bug fixes in the GSettings backend (crashers caused by use if
   custom dconf profiles)
 - some FreeBSD build fixes
 - increased Vala dependency to 0.11.4 (required for dconf-editor fixes)

Changes in dconf 0.6
====================

 - Rewrite a lot of the GSettings backend to reduce GDBus abuse.  We use
   our own worker thread now instead of trying to hijack GDBus's.
 - disable gobject-introspection support for now
 - drop support for GTK2 in dconf-editor
 - Add a new torture-test case
 - Increase dbus timeout to 2 minutes (in case the service is heavily loaded)
 - Fix several memory leaks and other bugs